+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [31-10-2024] 0.1.3+9
+- Add glob support in exclude-file
+- Add deep diving folder scan in assets
+
 ## [13-09-2024] 0.1.2+8
 - Increasig pub scores